+++ js/photos.js
@@ -1,5 +1,5 @@
 import { photos } from './data.js';
-import { onFullImageClick } from './full-photo.js';
+import {showFullImage} from './user-modul.js';
 
 const photoItemTemplate = document.querySelector('#picture')
   .content
@@ -14,7 +14,7 @@ const createPicture = (picture) => {
   photoElement.querySelector('.picture__img').src = picture.url;
   photoElement.querySelector('.picture__likes').textContent = picture.likes;
   photoElement.querySelector('.picture__comments').textContent = picture.comments.length;
-  onFullImageClick(photoElement, picture);
+  showFullImage(photoElement, picture);
   return photoElement;
 };
